client: use withRevision
